D:P: Fix usage of [[:^ascii:]]
Older perl versions do not support this syntax. In first case it is already
called only for Perl 5.6.0+ so just call it in eval to prevent compile
time warning for code which never runs. In second case it may be called
also for older Perl versions, so explicitly fallback to [^\x00-\x7F] when
older versions are used.